+++ src/komm/_error_control_block/decoders/majority_logic_repetition_code.py
@@ -8,7 +8,7 @@ from ..RepetitionCode import RepetitionCode
 def decode_majority_logic_repetition_code(
     code: RepetitionCode,
     r: npt.ArrayLike,
-) -> npt.NDArray[np.int_]:
+) -> npt.NDArray[np.integer]:
     u_hat = np.array([np.argmax(np.bincount(r))])
     return u_hat
